Add a new metadata file
hi DJ, this is a new metadata file describing the structure of the guide. With this file, it should be possible to build and publish the content with the edit button directly with the md content in the repo. I will test this as soon as this file is in the main branch. If it works, we will share with you the API call to that you can publish yourself.