check repo permission level, not team membership
Checking team membership would require two API calls, one for the
repo admin team, and one for the org level admin. This check allows
us to check the "outcome" of those combined team memberships, and
is actually what we need.

Also removed the first condition in the final step, as we shouldn't
need it; checking if disallowed is true should be enough.