replace MD001 rule with custom version
The output from DITA sometimes combines two markdown elements, putting a
heading (e.g. `###`) *inside of* a blockquote (`>`) like this:

```
> ### Some Heading
```

The standard Markdown linting, specifically rule
[MD001](https://github.com/DavidAnson/markdownlint/blob/main/doc/Rules.md#md001),
does not expect this and will raise an error when an error is not
wanted. Here is an example of where this has happened:

[Update Cloning_Repositories_7a68bfa.md Markdown linter

The approach here is to replace the standard MD001 with a custom one
(MD901) that avoids raising an error if the heading is combined with a
blockquote.