abort current function execution if blockquote
Having the blockquote check combined with the check for levels means that the error won't be emitted but the level calculation will still fire. This means that an actual erroneous heading level that follows a blockquoted heading may not be highlighted. So we want to jump out before that happens if we're in a blockquote-with-heading situation.